refactor: Move command context to utils
